refactor(controller): split up the workspace controller + easier to follow function names
